Added some documentation.
git-svn-id: https://midi-shape-shifter.googlecode.com/svn/trunk@74 4e31dc51-3177-9f86-f460-8117e2103d03

+++ MidiShapeShifter/MidiShapeShifter/Mss/MssEvent.cs
@@ -5,10 +5,16 @@ using System.Text;
 
 namespace MidiShapeShifter.Mss
 {
+    /// <summary>
+    ///     Represents an MssMsg that occured at a particular time.
+    /// </summary>
     public class MssEvent
     {
         public MssMsg mssMsg;
-        public double timestamp;
+        /// <summary>
+        ///     The system time that MssMsg occured in ticks
+        /// </summary>
+        public long timestamp;
 
         public MssEvent()
         {
